move functions from packager.packaging to packager.utils
