+++ packager/utils.py
@@ -20,6 +20,10 @@ import contextlib
 import os
 import shutil
 import tempfile
+from collections import namedtuple
+from xml.etree import ElementTree as ET
+from packager.textures import pack_textures
+import zipfile
 
 
 def makedirs_ignore_errors(path):
@@ -36,3 +40,66 @@ def tempdir():
         yield directory
     finally:
         shutil.rmtree(directory, ignore_errors=True)
+
+
+Artifact = namedtuple('Artifact', ['addon_id', 'version', 'git_repo', 'treeish'])
+
+
+def delete_companion_files(path):
+    for name in os.listdir(path):
+        # TODO: remove after krypton
+        if name.startswith("changelog-") and name.endswith(".txt"):
+            continue
+
+        if os.path.splitext(name)[1] != '.zip':
+            try:
+                if os.path.isdir(os.path.join(path, name)):
+                    shutil.rmtree(os.path.join(path, name))
+                else:
+                    os.remove(os.path.join(path, name))
+            except (IOError, OSError) as ex:
+                logger.warning("Failed to remove companion file '%s'" % os.path.join(path, name))
+                logger.exception(ex)
+
+
+def meets_version_requirements(imports, min_versions):
+    for addon_id, imported_version in imports:
+        if addon_id in min_versions:
+            if imported_version < min_versions[addon_id]:
+                return False
+    return True
+
+
+def pack_artifact(artifact, src_dir, dst_dir):
+    xml = ET.parse(os.path.join(src_dir, 'addon.xml'))
+    pack_textures(xml, src_dir)
+
+    # Copy asset files
+    assets = xml.find("./extension[@point='kodi.addon.metadata']/assets")
+    if assets is None:
+        assets = xml.find("./extension[@point='xbmc.addon.metadata']/assets")
+
+    if assets is not None:
+        for item in assets:
+            if item.text:
+                makedirs_ignore_errors(os.path.join(dst_dir, os.path.dirname(item.text)))
+                shutil.copyfile(os.path.join(src_dir, item.text), os.path.join(dst_dir, item.text))
+
+    else:  # for backwards compatibility with add-ons that do not use the assets element
+        if os.path.exists(os.path.join(src_dir, "icon.png")):
+            shutil.copyfile(os.path.join(src_dir, "icon.png"), os.path.join(dst_dir, "icon.png"))
+
+        if os.path.exists(os.path.join(src_dir, "fanart.jpg")):
+            shutil.copyfile(os.path.join(src_dir, "fanart.jpg"), os.path.join(dst_dir, "fanart.jpg"))
+
+        if os.path.exists(os.path.join(src_dir, "changelog.txt")):
+            shutil.copyfile(os.path.join(src_dir, "changelog.txt"), os.path.join(dst_dir, "changelog-%s.txt" % artifact.version))
+
+    # Write and compress files in src_dir to the final zip file
+    dest_file = os.path.join(dst_dir, "%s-%s.zip" % (artifact.addon_id, artifact.version))
+    with zipfile.ZipFile(dest_file, 'w', zipfile.ZIP_DEFLATED) as zf:
+        for root, dirs, files in os.walk(src_dir):
+            for name in files:
+                local_path = os.path.join(root, name)
+                archive_dest = artifact.addon_id + '/' + os.path.relpath(local_path, start=src_dir)
+                zf.write(local_path, archive_dest)
